requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.3
poppler-utils>=23.11.0

# Image Processing
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
easyocr>=1.7.1
numpy>=1.24.0

# Audio Processing
SpeechRecognition>=3.10.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Machine Learning & NLP
transformers>=4.35.0
torch>=2.1.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-huggingface>=0.0.6
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
sentence-transformers>=2.2.2

# Hugging Face
huggingface-hub>=0.19.0

# Audio Transcription
openai-whisper>=20231117

# Optional but recommended
tqdm>=4.66.1
requests>=2.31.0
